More WIP on slide adding

+++ src/frame.h
@@ -150,7 +150,7 @@ extern void insert_text_in_paragraph(Paragraph *para, size_t offs,
 extern void delete_text_in_paragraph(Paragraph *para,
                                      size_t offs1, size_t offs2);
 
-extern void split_paragraph(struct frame *fr, int pn, int pos,
+extern void split_paragraph(struct frame *fr, int pn, size_t pos,
                             PangoContext *pc);
 
 #endif	/* FRAME_H */
